Add functionality to support RTS/CTS test (Atheros)
When CTS_WIDTH is set to 40, the test case expects CTS to go only in 40
MHz. To achieve this we need to disable phy restart.

When DYN_BW_SGNL enable is triggered, RTS may be sent in any bandwidth
based upon the environment. The test case expects RTS to be in 80
bandwidth. This is achieved using wifitool internal command.

Signed-off-by: Jouni Malinen <jouni@qca.qualcomm.com>