Option to not add username/password as identity for EAP-SIM/AKA/AKA'
The new command line argument -A can be used to enable use of a real SIM
with EAP-SIM/AKA/AKA' by not configuring the identity and password
parameters from the command to enable simulated SIM (Milenage).

For eap_proxy based implementation, this new argument can be added to
allow wpa_supplicant to query the needed information from the Modem.

Signed-off-by: Jouni Malinen <jouni@qca.qualcomm.com>