OpenWrt: Detect and enable 'qcawifi' alone for Atheros drivers
As of now Atheros is the only OpenWrt driver subtype that needs special
support. If that subtype is selected, use "wifi detect" to enable only
the 'qcawifi' driver. This is needed to configure a single driver in
case of a three radio support which has 'mac80211' based driver and
qcawifi driver in a single image. In future we can provide a
configurable option for OpenWrt driver subtypes if the support for the
same is added.

Signed-off-by: Mohammed Shafi Shajakhan <mohammed@qti.qualcomm.com>