Use P2P virtual interface for P2P operations if needed
Implement support for virtual interface with P2P_DEVICE interface type.
This is used by wpa_supplicant for all P2P non-group operations
(discovery, GO negotiation, ..) with some drivers. It does not have an
associated netdev with it so it is simpler to implement. It is used to
separate scan and P2P search.

Signed-off-by: Jouni Malinen <jouni@qca.qualcomm.com>