traffic_agent: Derive Interface in traffic_agent_receive_start
traffic_agent_receive_start command is not sending the Interface
parameter but sigma_dut is expecting and using it to set power save in
Uapsd traffic packets. As the Interface is NULL, "iw" based power save
commands are failing leading to test case failures. Fix this by getting
interface from get_station_ifname() if the Interface parameter is not
available.

Signed-off-by: Jouni Malinen <jouni@qca.qualcomm.com>