sta_set_eaptls: Don't overwrite identity with "wifi-user@wifilabs.local"
sigma_dut used to overwrite the EAP identity which is being passed by
UCC command with "wifi-user@wifilabs.local" in EAP-TLS authentication.
New authentication servers do not require this static identity and may
fail authentication when it is used. So, don't overwite the identity
with "wifi-user@wifilabs.local" when UCC command includes the
username(identity) parameter.

Signed-off-by: Jouni Malinen <jouni@qca.qualcomm.com>