traffic_agent: Limit number of TX packets appropriately in send_file()
Earlier the sleep_time is being calculated only once when the
send_file() is called and if the sleep_time is 0 then, it is not
getting calculated as the packets are being transmitted. As a result
on certain platforms, where the packets are being transmitted rapidly
the total number of packets transmitted is exceeding by huge extent
the expected number of packets to be transmitted, which is equal
to (duration * frame_rate).

Signed-off-by: Jouni Malinen <jouni@qca.qualcomm.com>