Configure the WCN driver in the TDLS Explicit trigger mode
DRIVER_WCN drivers expect to be operational in the TDLS explicit trigger
mode which shall ensure a TDLS frame is transmitted when the
wpa_supplicant's control interface commands to initiate the TDLS
negotiation are executed. Hence, disable the TDLS external trigger
control mode (which enables the TDLS explicit trigger mode) in the host
driver and wpa_supplicant.

The following is the description of the three TDLS trigger modes.

TRIGGER_MODE_EXPLICIT: The trigger to initiate/teardown the TDLS
 connection to a respective peer comes from the user space.
 wpa_supplicant provides the commands TDLS_SETUP, TDLS_TEARDOWN,
 TDLS_DISCOVER to do this.
TRIGGER_MODE_IMPLICIT: Host driver triggers this TDLS setup/teardown
 to the eligible peer once the configured criteria (such as TX/RX
 threshold, RSSI) is met.
TRIGGER_MODE_EXTERNAL: Enables the driver to trigger
 the TDLS setup/teardown through the implicit mode only to the
 configured MAC addresses (wpa_supplicant,with tdls_external_control=1,
 configures the MAC address through TDLS_SETUP / TDLS_TEARDOWN commands).
 External mode works on top of the implicit mode. Thus the host driver
 is expected to configure in TDLS Implicit mode too to operate in
 External mode.

Signed-off-by: Jouni Malinen <jouni@qca.qualcomm.com>