Trigger "P2P_SET ps 0" on the same interface during sta_p2p_reset()
During the p2p_reset, the "P2P_SET ps 0" command is sent on the station
interface and the corresponding command to enable the power save
"P2P_SET ps 1" is sent on the respective interface (p2p). Thus, the
command to enable the power save does not reach to the driver, if the
power save state for the wdev interface (wdev->ps) in kernel is in
enabled state by default (CONFIG_CFG80211_DEFAULT_PS). This can be an
issue with drivers that do not have correctly synchronized power save
state with cfg80211 (cfg80211 assumes power save is enabled and tells
that to user space and drops requests to enable power save even though
the driver does not have power save enabled by default).

Work around this by also issuing "P2P_SET ps 0" on the same interface to
avoid need to fix the driver. The intention here is to not disturb the
existing implementation where this command is issued on the station
interface. This helps in not breaking the operability with the old host
drivers which expect this command on the station interface.

Signed-off-by: Jouni Malinen <jouni@qca.qualcomm.com>