cpu-boost: Rework scheduling setup
This patch minimizes the latency of an input boost by creating a dedicated
kworker and scheduling it with the realtime policy "SCHED_FIFO".
We use "MAX_RT_PRIO - 2" to have the input boost task preempt userspace
RT tasks if needed while being preempted by kernel internal RT tasks which are
scheduled at "MAX_RT_PRIO - 1" or higher.

Also since now the cpu-boost wq would only handle the work to disable a running
input boost, which isn't latency critical, we break up the cpu-boost wq and
move the work to disable a running input boost to the shared regular wq.

[ghostrider-reborn]: Applied to xiaomi powerkey boost as well

Signed-off-by: Alex Naidis <alex.naidis@linux.com>

cpu-boost: Bind to LITTLE cpus and reduce RT prio

To prevent the case of a possible RT priority inversion with
more important RT tasks like mdss (RIP), reduce the priority to 2.
This will make sure cpu-boost can take priority over most work
items while not causing a priority inversion leading to thread
interruption.

Additionally we bind the cpu-boost worker thread to CPUs 0-3 to
decrease the likeliness of it interfering with the app threads.
For that purpose, we do a small refactor of the kthread init process.

While at it, add logging for the case where sched_setscheduler fails,
as it can fail due to security checks.

Signed-off-by: Alex Naidis <alex.naidis@paranoidandroid.co>