dev_send_frame: Add BRP-RX and SLS frames for 60 GHz
Add support for the BRP-RX and SLS frames in the 60 GHz
program to dev_send_frame and implement using the
wil6210 driver. Support station and AP modes.
Both frames trigger a flow in the FW, so they are
implemented by sending a WMI command.
These commands can only be sent on a testbed machine.

Signed-off-by: Lior David <qca_liord@qca.qualcomm.com>