sta_reassoc: Issue scan before issuing reassoc (WCN)
The cld driver is not initiating a scan as part of FASTREASSOC
implementation and the command can fail if there is no scan entry for
the requested BSSID. Hence initiate a scan explicitly from sigma_dut.
Initiate a single channel scan if channel number is provided with
FASTREASSOC command else do a full scan.

Signed-off-by: Jouni Malinen <jouni@qca.qualcomm.com>