NAN: Fix to update subscribe id in subscribe followup
Subscribe id is not updated as part of subscribe follow up request,
which makes the follow up request to be sent out as new subscribe
request.

Fix the issue by updating the subscribe id from the earlier subscribe
request if the service name of subscribe request matches with the
current one.

Signed-off-by: Jouni Malinen <jouni@qca.qualcomm.com>