Revert "NAN: Fix NDP ping failure issue"
This reverts commit 93c6489a1407c5f0f072a25feadf013822b83dfb which
assumes that -I is deprecated (which it is not). On the other hand, some
devices may not support (%) which will lead to ping6 failure. Revert
this commit to get rid of the regressions. If that ping6 change for NAN
is needed in some cases, that needs to be redone in a safer way that
does not break ping6 for other users.

Signed-off-by: Jouni Malinen <jouni@qca.qualcomm.com>