Trigger FASTREASSOC to driver (WCN) only in connected state
sta_reassoc is also issued for the scenarios expecting the station to
reconnect to the AP with the maintained profile (cached PMKSA). Hence,
do not trigger FASTREASSOC to the driver in the disconnected state and
rather trigger REASSOCIATE command to wpa_supplicant. FASTREASSOC is
intended to trigger the roaming in the driver and thus is honored only
in the connected state.

Signed-off-by: Jouni Malinen <jouni@codeaurora.org>