Add support for sending netlink messages
This allows sigma_dut to issue nl80211 commands directly without having
to use an external tool like iw (or go through wpa_supplicant) to do so.

An example of how to send a vendor command to the driver is provided in
atheros.c by function cmd_atheros_config_scan(), which is to enable or
disable scan functionality in the driver.

Signed-off-by: Jouni Malinen <jouni@codeaurora.org>