DPP: Remove NetworkIntroResult,OK from SAE connections
wpa_supplicant will generate a PMKSA-CACHE-ADDED event for SAE as well
as DPP AKM, so it is not sufficient to use that alone to determine
whether DPP Network Introduction was used. Check the key_mgmt parameter
in the provisioned network profile to remove this extra information from
the SAE cases.

Signed-off-by: Jouni Malinen <jouni@codeaurora.org>