sta_send_addba: Return 1 (success) in case of no nl80211 support
Previously implementation to return a failure for no nl80211 support or
no vendor command support in the host driver would break the legacy
solution. Hence return 1 for such no support cases to match the earlier
behavior.

This is also aligned with the fallback behavior before the nl80211
support was introduced for this API. The comment in cmd_sta_send_addba()
seems to be clear that the explicit sending of ADDBA is not really
needed.

Signed-off-by: Jouni Malinen <jouni@codeaurora.org>