Enable client isolation for Hotspot 2.0 Single SSID cases
STAs associated with OSEN AKM must not be allowed to send frames to the
STAs associated with RSN(EAP). Enable client isolation for such AP
configuration for now to meet that requirement. It would also be
possible to handle this on per-STA basis by extending the
driver/mac80211 ap_isolate=1 support, but that's not yet available.

Signed-off-by: Jouni Malinen <jouni@codeaurora.org>