Update command handler process() callback return value type
The partial move from returning int to returning enum sigma_cmd_result
was problematic for LLVM/Clang Control Flow Integrity (CFI) sanitizer.
With CFI, the function pointers are verified to have the same type as
the actual callback function and enum is not considered to match int in
that context. This results in sigma_dut terminating if CFI is enabled
for the build. Avoid this by updating the process() callback functions
to use the exact same type as the function pointer.

This was done with following semantic patch and manual cleanup to remove
functions that matched but were not used as registered callbacks (those
functions can be addressed separately as cleanup since they do not cause
issues with CFI) and with coding style cleanup for indentation.

Signed-off-by: Jouni Malinen <jouni@codeaurora.org>