AP: Set IP config to the bridge for OWE Transition mode with LINUX-WCN
When AP OWE Transtion mode is tested with the "-b br-lan" command line
parameter, open and OWE BSSs are bridged into br-lan. Previously, IP
address was configured to one of the BSSs and not the bridge netdev, so
ping test to the other BSS failed. Fix this by configuring the IP
address into the bridge interface instead.

Signed-off-by: Jouni Malinen <jouni@codeaurora.org>