HS 2.0: Re-frame URL for special character (&)
hs20-action.sh on receiving events with URL, pass the string as-is to
e_loop. Since e_loop uses system() API to run the commands, it
recognizes & (ampersand) to run the process in background.

As an initial workaround for this, format the URL containing '&' by
replacing with escape character ("\\") before '&'. More robust solution
can be considered separately within e_loop to avoid use of system().

Signed-off-by: Jouni Malinen <jouni@codeaurora.org>