Allow separate 2.4 GHz and 5 GHz main/station ifnames to be set
This extends the -M and -S command line arguments to accept interface
name pair in <2.4 GHz ifname>/<5 GHz ifname> encoding to be set. For
example, "wlan0/wlan1". A new device_list_interfaces argument "band" can
then be used to specify which interface name is requested. This is an
initial step in allowing dynamic selection of a specific radio for
station operations based on the needed band.

Signed-off-by: Jouni Malinen <jouni@codeaurora.org>