Set DFS mode to enable in reset defaults for 802.11ax
Specifically for DUT cases, we need DFS scan to be enabled to make sure
in AUTO mode AP does not come up in DFS channel. For testbed case we
enable it as it is always good to do DFS scan although not a
requirement. This is similar to what was already done for VHT.

Signed-off-by: Priyadharshini Gowthaman <pgowtham@codeaurora.org>