sta_associate: Return status,COMPLETE if association does not complete
This replaces the previously used "status,ERROR,ErrorCode,Association
did not complete" with "status,COMPLETE,Result,Association did not
complete" to be more consistent with sta_associate behavior in error
cases. status,ERROR is now returned if the association attempt cannot be
started at all while status,COMPLETE is returned if wpa_supplicant has
started to process the request regardless of whether the association
succeeds. The "Result" argument is used with status,COMPLETE to provide
details on what happened with the attempt.

Signed-off-by: Jouni Malinen <jouni@codeaurora.org>