Use sae_confirm_immediate=2 to minimize latency
This moves to the optimized version of previously used
sae_confirm_immediate=1 approach in hostapd. The newer version allows
some key derivation steps to be removed from adding latency between the
operation. That change did not seem to have a significant difference in
timing, but it may actually be a noticeable improvement on some devices
where derivation of k (EC mul+add+mul) can actually take noticeable
amount of time.

Signed-off-by: Jouni Malinen <jouni@codeaurora.org>