Use a separate interface for NAN operations
This commit enhances the NAN behavior to use a separate interface
for the NAN discovery operations if such an interface is found.

This separate interface is expected to get created by the host driver
and the NAN operations use this interface if it exists. Otherwise, there
is a fallback to the "wlan0" interface to be backward compatible to the
implementations using it.

Signed-off-by: Jouni Malinen <jouni@codeaurora.org>