mdss: msm: hdmi: add CEC clear_logical_addr sysnode
When the logical address is cleared, CEC hal sets the logical
address to 15 but kernel cannot determine whether the logical
address is cleared or set to 15. To clearly detect that the
logical address is cleared, a new sysnode
"clear_logical_addr" is added.

If clear_logical_addr is set to 1, then kernel ignores
received CEC messages until it's reset to 0.

Change-Id: I517a3054c0816e774836767f3e0e47f92b395666
Signed-off-by: Venkata Prahlad Valluru <vvalluru@codeaurora.org>