msm: camera: cpas: Set AHB clock rate only if a client is streamed on
Currently the ahb clock rate is set as part of cpas_start for a
given client. In order to set this rate, any camera gdsc needs to
be enabled. This change adds a check to allow setting the clk rate
only if the gdsc regulator has been enabled to avoid NoC timeouts.

CRs-Fixed: 2671048
Change-Id: I6551c8c648ae298dca97f0e35ba7c9ca83b5a2cd
Signed-off-by: Karthik Anantha Ram <kartanan@codeaurora.org>