NAN: Fix the debug command data for NAN_TEST_MODE_CMD_TRANSPORT_IP_PARAM
The debug command data for NAN_TEST_MODE_CMD_TRANSPORT_IP_PARAM is
being overwritten with the structure data of NdpIpTransParams which
is incorrect, rather it should be passed in TLV format.

As part of commit 47d709ba866a ("NAN: Send IPv6 address, port and
protocol in TLV format"), the debug command data was constructed in TLV
format. However, there is an issue that before calling the
nan_debug_command_config(), the data is overwritten with the structure
data of NdpIPTansParam. This commit addresses that issue to update with
the correct TLV data.

Signed-off-by: Jouni Malinen <jouni@codeaurora.org>