Use the tos=46 for user_priority=6 and Voice traffic class
RFC 8325 updates the DSCP to UP mappings, specifically maps the
current ToS value of 48 to user_priority 7.

Both the current implementations and RFC 8325 map the ToS 46 to
user_priority 6.

Thus, use the ToS value of 46 for the user_priority 6 to suite the
current implementation and the update proposed in RFC8325.

Signed-off-by: Jouni Malinen <jouni@codeaurora.org>