msm: camera: ife: Detect RDI only context before acquire hw
In current code, rdi only context is upated after acquiring the
hw. In cases, where the first in_port comes as RDI only and second
in_port comes as Pix and RDI, IFE lite along with Full IFE will
be acquired. Context in question will not be marked as rdi_only.
This causes issues as the interrupts for Lite will be enabled only
if the context is rdi_only. Also, the rdi_only state machine will
not be activated.
To handle such scenarios, before acquiring the hw, we now detect
whether the context is rdi_only. Based on this, hw manager can
pass the information to CSID driver if LITE can be used or not.
Based on this, the reserve call can take decision inside CSID
driver. This information can also be used in advanced targets
where LITE has more capabilities and CSID driver needs more
information to make decisions while reserving hw.
This commit fixes some existing issues of releasing the
resources while handling the early exit from acquire call. It also
adds logging in case there is a failure in acquire because of non
availability of required resources.

CRs-Fixed: 2794147
Change-Id: I0591a83c23504eb67f3b87f17e172a9c67e7c82b
Signed-off-by: Gaurav Jindal <gjindal@codeaurora.org>