AP: Enable OWE PTK workaround when a specific group is not set
OWE PTK workaround needs to be enabled for all cases when configuring
OWE instead of only the cases that set a specific owe_group value. This
fixes an issue in cases where ECGroupID is not set, but the workaround
is needed as legacy STAs can choose group greater than 19 for the
connection.

Signed-off-by: Veerendranath Jakkam <vjakkam@codeaurora.org>