ARM: dts: msm: Introduce demura HFC region nodes
ABL expects the demura HFC region nodes to always be present
under a device's reserved memory regions. If this node is not
present, ABL is unable to insert it and demura continuous
splash cannot be handled.

Introduce dummy demura HFC regions into the device tree.
These nodes are given an address and size of 0 to allow the
kernel to specially handle them.

Change-Id: I6b95a145b4d515e24a454d9ad293e477e6dec66f