ARM: dts: msm: update interconnect and global interuupts on yupik
This change does the following:

1) Remove kgsl smmu interconnect node. For TCU config register access we
shouldn't required DDR path votes.  KGSL smmu already votes for clock
and regulator.

2) Remove the tcu_comb_irpt_ns for both apps and kgsl smmu as the
interrupt is OR of context interrupts, TBU perf interrupts and global
fault interrupts. This interrupt is not really required and registering
for the same can showup as the global faults in EL1.

Change-Id: I4de4a866d936f1f7f053479dc00c72c46e84d80d