Makefile: Optimize compiler specific instructions for coupled A76.A55
On an ARM platform(unlike x86), -march and -mtune flags are entirely
orthogonal. This means, -march and -mtune does not over-ride each other!

This means, using -march=X -mtune=Y tells the compiler to generate
binary code for architecture X and to tune it for microarchitecture Y.
The resulting binary will execute on architecture X and all supersets
of architecture X, but will be optimized for microarchitecture Y.

-mcpu=Z flag, on ARM, simultaneously determines the architecture plus its
extensions and the microarchitecture too.  The permissible values of -mcpu
are same as those available for -mtune. This complies, on an ARM platform,
using -mcpu=Z determines what kind of instructions it can emit while generating
assembly codes for the specified architecture (synonymous to -march=X) and to
determine the target processor for which to tune the codes for (synonymous to -mtune=Y).

GCC supports tuning for big.LITTLE system, while Clang / LLVM does not
at these point. On LLVM, optimizing for the LITTLE cluster is considered
and the compiler automatically does this even if we set flags for
tuning big.LITTLE .

-mcpu, -march and -mtune flag's usage has been nicely documented in

[1] https://community.arm.com/developer/tools-software/tools/b/tools-software-ides-blog/posts/compiler-flags-across-architectures-march-mtune-and-mcpu
[2] https://gcc.gnu.org/onlinedocs/gcc/AArch64-Options.html
[3] https://en.wikichip.org/wiki/arm_holdings/microarchitectures/cortex-a76

Change-Id: I647cc78b24039bf059a1a516a99150935fd53199
Signed-off-by: Panchajanya1999 <panchajanya@azure-dev.live>
(cherry picked from commit 3bc76811d431c589d6d991026f93fc8fc5a6c824)
Signed-off-by: psndna88 <psndna88@gmail.com>

# Conflicts:
#	Makefile