ARM: dts: msm: Reorganize the apps_rsc and mxlvl node for sa8195
Move the 'apps_rsc' node to the beginning of the 'soc'
node such that the rpmh-regulator devices are populated
early in the boot process. As many consumers' probe depend
on these regulator devices, placing 'apps_rsc' at the beginning
would satisfy the consumer devices as soon as they are
populated later in the boot. This way, the drivers no longer
have to wait for the regulator devices to be populated, thus
optimizing the boot performance.

Change-Id: I8f760a8e3eb3c0e301e5866d83e828208743c183