msm: mhi-dev: Block the adb read when mhi is in suspend
Adb client on device expects a blocking read from mhi. While mhi
is in suspend, it is returning with an error value. This is
causing the adb client to spawn every 1ms causing error code to
flood the serial logs. If mhi is in suspend and adb client tries
to poll for data, remove the read request from the list to block
the read.

Change-Id: Idaa70568cbddaf6b99fb5c6722a4da594d3e066d
Signed-off-by: Gauri Joshi <gaurjosh@codeaurora.org>