dt-bindings: Add optional EUD parameter for interrupt mask
The EUD driver defaults to using INT1 as the interrupt mask
for the apps processor. However, on modem, this interrupt
does not serve the apps processor, so add a new optional
device tree binding which can specify the offset of the
interrupt mask used for the apps processor and override the
default.

Change-Id: Id73336dcf737945effc254d3179bbfe354a5026b