ARM: dts: msm: add dummy regulator for ibb supply for Blair
When the AMOLED panel is powered from the external power source,
the IBB regulator is not needed. In the absence of ibb regulator
node we are getting error message (non-fatal) for AMOLED panel's
operation on ibb during change in the power states. Using stub
regulator for ibb supply to fix this error. The Blair display DT 
includes DT of PM6125 variant of Holi, so this change applies to 
both.

Change-Id: I851cef3e6c235b8ee2fa334c45c96caa038262bf