sta_scan: Support the RxMac parameter in the scan command
Add support to parse RxMac in the sta_scan command to configure the
driver to use scan request BSSID value in Probe Request frame RA(A1) for
scan. If this configuration is set, the driver uses the BSSID value from
the scan request to set the RA(A1) in the Probe Request frames during
the scan, else the broadcast address is set in the Probe Request frames
RA(A1).

Signed-off-by: Jouni Malinen <jouni@codeaurora.org>