AP: Enable MU beamformer capability based on driver capabilities
MU beamformer capability was enabled by default for HE program from
ap_reset_default based on the driver. Due to this APUT bringup could
fail for the hardware modes which are not MU beamformer capable in AP
mode like IPQ5018. So, disable MU beamformer configuration if the driver
does not indicate support for MU beamformer capability.

Co-developed-by: Sowmiya Sree Elavalagan <ssreeela@codeaurora.org>
Signed-off-by: Sowmiya Sree Elavalagan <ssreeela@codeaurora.org>
Signed-off-by: Venkateswara Naralasetty <vnaralas@codeaurora.org>