QM: Add support to configure DSCP policies using iptables
Add support to configure DSCP policy rules using iptables.
This is needed since Android doesn't have support for nftables.

This commit by default uses iptable rules for Android. While the current
implementation does not enable this for non-Android cases, it is
possible that the current design would be extended to cover some other
cases as well and as such, only the selection step is within a platform
specific compilation block.

Signed-off-by: Shivani Baranwal <quic_shivbara@quicinc.com>