ANDROID: restore some more removed refcount functions
Due to the refcount functions being cleaned up and majorly simplified in
5.4.208, a number of previously exported functions that were part of
inline functions were removed and the real functions used instead.  As
those were part of the Android ABI, restore functions for these
"checked" refcount abis so that existing code continues to build
properly, while anything that is rebuilt, will be able to take advantage
of the new inline functions instead.

The functions restored are:
	refcount_add_not_zero_checked
	refcount_add_checked
	refcount_sub_and_test_checked

Fixes: d0d583484d2e ("locking/refcount: Consolidate implementations of refcount_t")
Signed-off-by: psndna88 <psndna88@gmail.com>